Remove net from dependencies
