+++ components/script/Cargo.toml
@@ -76,7 +76,6 @@ mime = "0.3.13"
 mime_guess = "2.0.0"
 mitochondria = "1.1.2"
 msg = { path = "../msg" }
-net = { path = "../net" }
 net_traits = { path = "../net_traits" }
 num-traits = "0.2"
 parking_lot = "0.11"
